common/kol_session/pyproject.toml
[project]
name = "kol-session"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "python-dotenv>=1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["kol_session"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
